use dlopen() so we don't need to compile
For now, let's try and avoid the compile step because
it's one less thing we'll need to do in setup.py.  This
also makes it easier on a consumers and developers of Twisted
because there's less to setup on your local system just to run
the tests.

This change also removes the scratch file with some older code
in it.